support trimming

+++ play.sh
@@ -1,25 +1,40 @@
 #!/bin/bash
 
-if [ $# -lt 2 ] ; then
-  echo "Usage: ./play.sh [user@]host:<wavs_dir> <wavlist> [start_index]"
+if [ $# -lt 1 ] ; then
+  echo "Usage: ./play.sh [user@]host:<wav_list> [start_index]"
   exit
 fi
 
-IFS=$'\n' read -d '' -r -a wavs < $2
 i=0
-if [ $# -eq 3 ] ; then
-  i=$3
+if [ $# -eq 2 ] ; then
+  i=$2
 fi
 
+scp $1 wav.lst
+IFS=$'\n' read -d '' -r -a wavs < wav.lst
 while [ $i -lt ${#wavs[@]} ] ; do
-  wav=${wavs[$i]}
-  scp $1/$wav .
-  sox $wav -d
-  rm $wav
+  IFS=' ' read -r -a array <<< ${wavs[$i]}
+  wav="audio.wav"
+  [ -e $wav ] && rm $wav
+  scp ${1%:*}:${array[0]} $wav
   echo "=================================================================="
-  echo "Played ($((i+1))/${#wavs[@]}): $wav."
+  echo "($((i+1))/${#wavs[@]}): ${array[0]}"
+  if [ ${#array[@]} -eq 1 ] ; then
+    echo "=================================================================="
+    sox $wav -d
+  elif [ ${#array[@]} -eq 2 ] ; then
+    echo ${array[1]}
+    echo "=================================================================="
+    sox $wav -d
+  elif [ ${#array[@]} -eq 3 ] ; then
+    echo "=================================================================="
+    sox $wav -d trim ${array[1]} ${array[2]}
+  elif [ ${#array[@]} -eq 4 ] ; then
+    echo ${array[3]}
+    echo "=================================================================="
+    sox $wav -d trim ${array[1]} ${array[2]}
+  fi
   echo "Press [r]eplay [n]ext [p]revious [q]uit"
-  echo "=================================================================="
   read -rsn1 key
   if [ "$key" = "q" ] ; then
     exit
@@ -27,13 +42,13 @@ while [ $i -lt ${#wavs[@]} ] ; do
     i=$((i+1))
     if [ $i -ge ${#wavs[@]} ]; then
       let i=${#wavs[@]}-1
-      echo "=======================Reached end of list.======================="
+      echo "!!!!Reached end of list!!!!"
     fi
   elif [ "$key" = "p" ] ; then
     i=$((i-1))
     if [ $i -lt 0 ]; then
       i=0
-      echo "====================Reached beginning of list.===================="
+      echo "!!!!Reached beginning of list!!!!"
     fi
   elif [ "$key" != "r" ] ; then
     echo "Unknown key: $key"
